% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctsGE.R
\name{ClustIndexes}
\alias{ClustIndexes}
\title{Clustering the indexes applying kmeans}
\usage{
ClustIndexes(x, scaling = TRUE)
}
\arguments{
\item{x}{A ctsGEList object}

\item{scaling}{Boolean parameter, does the data should be standardized
before clustered. Default = TRUE}
}
\value{
ctsGEList object is returned as output,
with the relative culstered indexes table in
 \emph{object$ClusteredIdxTable},
and the number of clusters for each index in \emph{object$optimalK}
}
\description{
Clustering each index, that was predifined by
\code{\link{PreparingTheIndexes}}, with \code{\link{kmeans}}.
}
\details{
In order to choose a "good" k for K-means clustering,
we apply the Elbow method:
First of all, compute the sum of squared error (SSE) for different
values of k (for example 1-10).
The plot of k against the SSE shows that the error decreases as
k gets larger;
The idea of the elbow method is to choose the k at which the SSE
decreases abruptly.
In order to detect the right k where the SSE decreases abruptly,
we compute ratio of the within cluster sum of squares(WSS=SSE) to the total
sum of squares(TSS),and assuming that ratio of less than 0.2 indicate that
from that point error doesn't decrease much and adding another cluster
doesn't give much better modeling of the data.

By default data is standardize before clustering,
for clustering the raw counts set the \command{\strong{scaling}}
parameter to
 FALSE.
}
\examples{

data_dir <- system.file("extdata", package = "ctsGE")
files <- dir(path=data_dir,pattern = "\\\\.xls$")
rts <- readTSGE(files, path = data_dir,
labels = c("0h","6h","12h","24h","48h","72h"), skip = 10625 )
prts <- PreparingTheIndexes(rts)

tsCI <- ClustIndexes(prts)

head(tsCI$ClusteredIdxTable) #the table with the clusterd indexes
head(tsCI$optimalK) #the table with the number of clusters for each index

}
\seealso{
\code{\link{kmeans}}, \code{\link{PlotIndexesClust}}
}

