% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctsGE.R
\name{ClustProfiles}
\alias{ClustProfiles}
\title{Clustering the profiles}
\usage{
ClustProfiles(x, scaling = TRUE)
}
\arguments{
\item{x}{A ctsGEList object}

\item{scaling}{Boolean parameter, does the data should be standardized before clustered. Default = TRUE}
}
\value{
ctsGEList object is returned as output, with the relative culstered profiles table in \emph{object$ClusteredProfilesTable}, and the number of clusters for each profile in \emph{object$optimalK}
}
\description{
Clustering each profile, that was predifined by \code{\link{PreparingTheProfiles}}, with \code{\link{kmeans}}.
}
\details{
In order to choose the right k(number of clusters), so that the clusters are compact and well separated, we compute ratio of the within cluster sum of squares(WSS) to the total sum of squares(TSS).

when: \eqn{WSS / TSS < 0.2} this is the optipal k

It is often recommanded to standardized the data before clustering, the standardize method is the one you chose in \code{\link{PreparingTheProfiles}}.

If you are not intresting in standardizing your data before clustering, you can set the \command{\strong{scaling}} parameter to FALSE.

If you want to cluster a specific profile, you do not have to cluster all your profiles, you can skip this step and directly run \emph{PlotProfilesClust}
with the profile of your choise, for more details see: \code{\link{PlotProfilesClust}}
}
\examples{
\dontrun{

 tsCP <- ClustProfiles(PRTS)
 head(tsPS$ClusteredProfilesTable) #the table with the clusterd profiles
 head(tsPS$optimalK) #the table with the number of clusters for each profile
}
}
\seealso{
\code{\link{scale}},  \code{\link{kmeans}}, \code{\link{PlotProfilesClust}}
}

