% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctsGE.R
\name{MadScale}
\alias{MadScale}
\title{Scaling by the Median Absolute Deviation}
\usage{
MadScale(x)
}
\arguments{
\item{x}{A numeric}
}
\value{
The MadScale value of the input
}
\description{
MadScale is a function whose default centers and scales the rows of a numeric matrix,
this scaling is made for the purpse of time series profiling. for more information see details
}
\details{
Centering is done by subtracting the row median.

Scaling is done by dividing the (centered) rows of x by their \code{\link{mad}} (median standard deviations)


\tabular{r}{
\deqn{x - median(x)/mad(x))}\cr}
}
\examples{
(rawCounts <- matrix(c(103.5, 75.1, 97.3, 27.12, 34.83, 35.53, 40.59, 30.84, 16.39, 29.29),2,5))

(msCounts <-MadScale(rawCounts))
}
\seealso{
\code{\link{scale}},  \code{\link{mad}}
}

