% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctsGE.R
\name{PlotIndexesClust}
\alias{PlotIndexesClust}
\title{A graphic visualization of an index}
\usage{
PlotIndexesClust(x, idx, k = NULL, scaling = TRUE)
}
\arguments{
\item{x}{ctsGEList object}

\item{idx}{A character, the index to plot
(e.g., for 8 time points "11100-1-1-1")}

\item{k}{A numeric, number of clusters. If not given the function will
calculate what is the optimal k for the index.}

\item{scaling}{A boolean, default to TRUE,
does the data should be standardized before clustered with kmeans.}
}
\value{
A list with two objects:
 \enumerate{\item Table of the index and its clusters
            \item Gene expression pattern graphs
            for each one of the clusters}
}
\description{
Line graphs made with \code{\link{ggplot}} of a index and its clusters -
the user choose the index to show,
clustering is done with \code{\link{kmeans}}
}
\examples{

data_dir <- system.file("extdata", package = "ctsGE")
files <- dir(path=data_dir,pattern = "\\\\.xls$")
rts <- readTSGE(files, path = data_dir,
labels = c("0h","6h","12h","24h","48h","72h"), skip = 10625 )
prts <- PreparingTheIndexes(rts)
pp <- PlotIndexesClust(prts,idx="000011")
pp$graphs # plots the line graphs

}
\seealso{
\code{\link{ggplot}},  \code{\link{kmeans}},
 \code{\link{ClustIndexes}}
}

