% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctsGE.R
\name{PlotProfilesClust}
\alias{PlotProfilesClust}
\title{A graphic visualization of a profile}
\usage{
PlotProfilesClust(x, idx, k = NULL, scaling = TRUE)
}
\arguments{
\item{x}{ctsGEList object}

\item{idx}{A character, the profile to plot (e.g., for 8 time points "11100-1-1-1")}

\item{k}{A numeric, number of clusters. If not given the function will calculate what is the optimal k for the profile.}

\item{scaling}{A boolean, default to TRUE, does the data should be standardized before clustered.}
}
\value{
A list with two objects: \enumerate{\item Table of the profile and its clusters
                                            \item Gene expression pattern graphs for each one of the clusters}
}
\description{
Line graphs made with \code{\link{ggplot}} of a profile and its clusters - the user choose the profile to show,
clustering is done with \code{\link{kmeans}}
}
\examples{
\dontrun{

 pp <- PlotProfilesClust(PRTS,idx="000011") # when 'k' is not provided, the function will determine the number of clusters
 pp$graphs # plots the line graphs
 pp$Clust_k_00011 # the table of genes in the profile seperated to clusters
}
}
\seealso{
\code{\link{ggplot}},  \code{\link{kmeans}}, \code{\link{ClustProfiles}}
}

