% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctsGE.R
\name{PreparingTheIndexes}
\alias{PreparingTheIndexes}
\title{Define an expression index for each gene}
\usage{
PreparingTheIndexes(x, cutoff = 1, mad.scale = TRUE)
}
\arguments{
\item{x}{A ctsGEList object}

\item{cutoff}{A numeric that define the degree of change in gene
espression rate.
 \emph{See Details}.}

\item{mad.scale}{A boolean defaulting to TRUE as to what method of
scaling to use.
Default median-base scaling. FALSE, mean-base scaling.}
}
\value{
ctsGEList object is returned as output with the relative
standarization table in \emph{object$scaled},
and the indexes table in \emph{object$index}
}
\description{
Reads the table of genes expression and return an expression index for each
gene.
}
\details{
After the standardize step, there is an indexing step to convert
each time point to \strong{0, 1, -1}:

 The function defines limits around the center, median or mean, +/- cutoff
 value in median absolute deviation(mad) or standard deviation(sd) units
 respectively.
 The signification level limit is determined by the user with the
 \command{cutoff} option.

 Then the function calculate a value to each time point according to:
\enumerate{
        \item \bold{0}  when gene expression at a specific time point is
        within the center limits (+/- cutoff).
        \item \bold{1}  when gene expression at a specific time point is
        greater than center top limit (+ cutoff).
        \item \bold{-1} when gene expression at a specific time point is
        smaller than center bottom limit (- cutoff).}
}
\examples{
data_dir <- system.file("extdata", package = "ctsGE")
files <- dir(path=data_dir,pattern = "\\\\.xls$")
rts <- readTSGE(files, path = data_dir,
labels = c("0h","6h","12h","24h","48h","72h"), skip = 10625 )
prts <- PreparingTheIndexes(rts)

}
\seealso{
\code{\link{scale}} \code{\link{index}}
}

