% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctsGE.R
\name{PreparingTheProfiles}
\alias{PreparingTheProfiles}
\title{Predefine set of profiles}
\usage{
PreparingTheProfiles(x, cutoff = 1, mad.scale = TRUE)
}
\arguments{
\item{x}{A ctsGEList object}

\item{cutoff}{A numeric that define the degree of change in gene espression rate. \emph{See Details}.}

\item{mad.scale}{A boolean defaulting to TRUE as to whether or not to use \code{\link{MadScale}} for standardize the data, choose FALSE for standardization with \code{\link{scale}} function.}
}
\value{
ctsGEList object is returned as output with the relative standarization table in \emph{object$scaled}, and the profiles table in \emph{object$profiles}
}
\description{
Reads the table of genes expression and return a set of model expression profiles,
that representive of any gene expression pattern.
}
\details{
The raw expression data values are first standardized with \code{\link{MadScale}}

After the scaling step, there is an indexing step to convert each time point to \strong{0 / 1 / -1}:

\command{cutoff} - Define the degree of change in gene expression level throughout the given time points.

The user can decide what is the degree, in which the expression level, is not significant enough or not informative enough for their experiment purpose.

 For example: if cutoff is set to 0.7, genes that their standardized value in a certain time point is:
\enumerate{
       \item between 0.7 to -0.7 will get the index \bold{0} - no significant change has made in gene expression at that time-point
       \item greater than 0.7 will get the index \bold{1} - gene expression was up regulated at that time-point
       \item smaller than -0.7 will get the index  \bold{-1} - gene expression was down regulated at that time-point}

 Please see example in \code{\link{index}}, on how different cutoffs produce different expression patterns,
 we highly recommand to test different \bold{cutoffs} in order to set the one that suits your data the most.
}
\examples{
\dontrun{
files <- dir(pattern="*\\\\.txt$")
RTS <- readTSGE(files)
PRTS$tsTable # the gene expression raw count table
PRTS <- PreparingTheProfiles(RTS)
PRTS$scaled # the gene expression standardized count table
PRTS$profiles # the profiles of gene expression
}

}
\seealso{
\code{\link{scale}}, \code{\link{MadScale}}, \code{\link{index}}
}

