% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctsGE.R
\name{index}
\alias{index}
\title{Indexing step for Time Series Supervised Clustering}
\usage{
index(x, cutoff = 1)
}
\arguments{
\item{x}{A numeric}

\item{cutoff}{A numeric, dermine the threshold for indexing, values between \emph{cutoff and -cutoff} will get the value \bold{0}, meaning no significant change was made in gene expression. Default = 1}
}
\value{
Gene expression profile
}
\description{
Takes a numeric vector and return a profile vector that represent the gene expression pattern
}
\examples{
rawCounts <- c(103.5, 75.1, 97.3, 27.12, 34.83, 35.53, 40.59, 30.84, 16.39, 29.29)

(msCounts <-MadScale(rawCounts)) # standardized step with MadScale

cutoff <- seq(0.2,2,0.1)    # different cutoff produce different profiles

for(i in cutoff){print(index(msCounts,i))}
}
\seealso{
\code{\link{PreparingTheProfiles}}
}

